third_party/blink/web_tests/virtual/mutation-events-enabled/fast/events/mutation-events-disabled-expected.txt

CONSOLE WARNING: Listener added for a synchronous 'DOMSubtreeModified' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMSubtreeModified' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMNodeInserted' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMNodeInserted' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMNodeRemoved' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMNodeRemoved' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMNodeRemovedFromDocument' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMNodeRemovedFromDocument' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMNodeInsertedIntoDocument' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMNodeInsertedIntoDocument' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMCharacterDataModified' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMCharacterDataModified' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMNodeRemoved event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMNodeRemoved event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMNodeRemovedFromDocument event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMSubtreeModified event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMNodeInserted event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMNodeInserted event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMNodeInsertedIntoDocument event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMSubtreeModified event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMSubtreeModified event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMSubtreeModified event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMCharacterDataModified event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMCharacterDataModified event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMSubtreeModified event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMSubtreeModified event Reached unreachable code
CONSOLE WARNING: Listener added for a synchronous 'DOMSubtreeModified' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMNodeInserted' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMNodeRemoved' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMNodeRemovedFromDocument' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMNodeInsertedIntoDocument' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMCharacterDataModified' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught Error: assert_unreached: Received DOMSubtreeModified event Reached unreachable code
Harness: the test ran to completion.
